requirements.txt
requests==2.21.0
cwltool==3.0.20200720165847
cwlgen==0.4.0
cwl-utils==0.7 #just make the parser command_line_tool.py from scratch, however
schema-salad==7.0.20200811075006
ruamel.yaml
semantic-version
miniwdl